requirements.txt
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
# https://github.com/omarryhan/aiogoogle
aiogoogle==5.3.0
